project_jarvis/project_jarvis/data_engineering/requirements.txt
azure_storage==0.37.0
azure-identity==1.15.0
azure-search-documents==11.6.0b1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
langchain==0.1.11
langchain_community==0.0.27
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
Pillow==10.2.0
pypdf==4.1.0
python-dotenv==1.0.1
tenacity==8.2.3
tiktoken==0.6.0
typing_extensions==4.10.0
unstructured==0.12.5
poetry ==1.8.2
python-magic-bin==0.4.14
